melhorando estrutura da interface
