+++ interface/app/controllers/utils.py
@@ -1,0 +1,6 @@
+from app import app
+
+
+def allowed_file(filename):
+    return '.' in filename and \
+           filename.rsplit('.', 1)[1].lower() in app.config['ALLOWED_EXTENSIONS']
