+++ interface/config.py
@@ -3,4 +3,5 @@ import os
 DEBUG = True
 UPLOAD_FOLDER = '/home/pediogo/Documentos/pastaDosUploads'
 PATH_GPON_PORTAS = '/home/pediogo/Documentos/broadband_internet_analysis/gpon/ports'
+ALLOWED_EXTENSIONS = ['csv', 'xlsx']
 SECRET_KEY = os.urandom(16)
