+++ interface/run.py
@@ -1,4 +1,4 @@
 from app import app
 
 if __name__ == '__main__':
-    app.run()
+    app.run(host='0.0.0.0')
